Function zero2prod::telemetry::get_subscriber

pub fn get_subscriber(
    name: String, 
    env_filter: String
) -> impl Subscriber + Send + Sync

Compose multiple layers into a tracing's subscriber.

Implementation Notes

We are using impl Subscriber as return type to avoid having to spell out the actual type of the returned subscriber, which is indeed quite complex. We need to explicitly call out that the returned subscriber is Send and Sync to make it possible to pass it to init_subscriber later on.